Moved Box out when Dropped and Allowed it to be on Side Player is Facing
Moves the box out of the way of the player collider when dropped, by
pushing it out in front of the direction they are facing.  Also it
attaches itself to the player on the side that the player is facing when
they grabbed it, currently working on updating it when the player turns,
may be able to do just by rotating the player.